Move unit tests down a folder
